requirements.txt
langchain==0.3.12
google-generativeai==0.8.3
langchain-google-genai==2.0.7
chromadb==0.5.23
pypdf2==3.0.1
python-dotenv==1.0.1
PyPDF==5.1.0
langchain-community==0.3.12
langchainhub==0.1.21
cohere==5.13.3
ipywidgets==8.1.5
streamlit==1.41.1
pysqlite3-binary